requirements.txt
requests
beautifulsoup4
PyYAML
line-bot-sdk
